% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rice_index.R
\name{calc_rice_index}
\alias{calc_rice_index}
\title{Calculating the Rice Index}
\usage{
calc_rice_index(votecounts)
}
\arguments{
\item{votecounts}{A vector containing the number of votes for yes and no in any given order. If a vector of length 1 is given, it is assumed that all votes were casted for the same option.}
}
\description{
This function allows you to easily calculate the Rice Index for a given vote.
}
\examples{
Examples of using calc_rice_index():

test_rice <- c(15,2)

test_rice_shorter <- c(15)

test_rice_longer <- c(15,2,5)

calc_rice_index(test_rice)

calc_rice_index(test_rice_shorter)

calc_rice_index(test_rice_longer)
}
\keyword{rollcall}
