% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enumerate.R
\name{enumerate}
\alias{enumerate}
\title{Enumerate contents of a vector to a single string}
\usage{
enumerate(itemvector, german = F, oxford = F)
}
\arguments{
\item{german}{Default==F If set to True enumerate() prints "und" instead of "and".}

\item{oxford}{Default==F If set to True enumerate() includes an oxford comma.}

\item{itemsvector}{Your vector to be enumerated.}
}
\description{
This function allows you to easily enumerate content of a vector to one long string, including an "and" at the end.
}
\examples{
Examples of using enumerate():

examplevector <- c("Germany", "Italy", "Spain")

enumerate(examplevector)

enumerate(examplevector, german=T, oxford=T)
}
\keyword{string}
\keyword{strings}
