% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_wbt_parlgov.R
\name{plot_wbt_parlgov}
\alias{plot_wbt_parlgov}
\title{Plotting Turnout from ParlGov Data}
\usage{
plot_wbt_parlgov(countryname, german = F)
}
\arguments{
\item{countryname}{A vector of strings containing the english names of countries to be plotted.}

\item{german}{Default=False. If True text in the plot will be printed in German.}
}
\description{
This function allows you to easily plot election turnout for european countries. It plots turnout at elections for the national parliament as well as for the european parliaments if available. Data is sourced from the great ParlGov Dataset: Döring, Holger and Philip Manow. 2019. Parliaments and governments database (ParlGov): Information on parties, elections and cabinets in modern democracies. Development version. URL: http://www.parlgov.org/
}
\examples{

plot_wbt_parlgov("Spain")

plot_wbt_parlgov(c("Poland","Germany"), german=T)
}
\keyword{elections}
\keyword{turnout}
