% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_plot_template.R
\name{save_plot_template}
\alias{save_plot_template}
\title{Exporting Plots for use in LaTeX}
\usage{
save_plot_template(
  plot_object = ggplot2::last_plot(),
  filenames,
  pdfwidth,
  pdfheight,
  latexwidth = 1,
  latexlabel,
  latexcaption
)
}
\arguments{
\item{plot_object}{Plot to save. Defaults to last_plot()}

\item{filenames}{name of the .tex and .pdf files}

\item{pdfwidth}{Width of the .pdf file in inches}

\item{pdfheight}{Height of the .pdf file in inches}

\item{latexwidth}{Proportion of the linewidth to use in LaTeX. Defaults to 1}

\item{latexlabel}{LaTeX label for referencing the plot in the LaTeX document}

\item{latexcaption}{Figure Caption in LaTeX}
}
\description{
Exporting Plots for use in LaTeX
}
\keyword{string}
\keyword{strings}
\keyword{utility}
